+++ scripts_matlab/old_scripts/GFP_AUC_script_ez_UI_imadjust.m
