+++ scripts_matlab/old_scripts/GFP_AUC_script_testing.m
