+++ scripts_matlab/old_scripts/false_sharpen.m
