+++ scripts_matlab/old_scripts/find_adjacent_pixel.m
