+++ scripts_matlab/old_scripts/mean2mask.m
