+++ scripts_matlab/old_scripts/single_sample/bad_images_manual_fix.m
