+++ scripts_matlab/old_scripts/straighten_worms.m
