+++ scripts_matlab/testing/GFP_AUC_script_ez_UI_dual_excitation2.m
