+++ scripts_matlab/testing/convert_sicko_to_lightsaver_otherimgs.py
