+++ scripts_matlab/testing/find_adjacent_pixel.m
