+++ scripts_matlab/testing/straighten_worms.m
